+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/list/MessageItem.kt
@@ -459,7 +459,7 @@ public fun DefaultMessageItemLeadingContent(
 @Composable
 public fun DefaultMessageItemHeaderContent(
     messageItem: MessageItem,
-    modifier: Modifier,
+    modifier: Modifier = Modifier,
 ) {
     val message = messageItem.message
 
@@ -569,20 +569,16 @@ public fun DefaultMessageItemContent(
     onGiphyActionClick: (GiphyAction) -> Unit = {},
     onImagePreviewResult: (ImagePreviewResult?) -> Unit = {},
 ) {
-    val (message, position, parentMessageId, ownsMessage, _) = messageItem
+    val (message, position, _, ownsMessage, _) = messageItem
 
-    val bubbleShape = if (message.id == parentMessageId) {
-        ChatTheme.shapes.myMessageBubble
-    } else {
-        when (position) {
-            Top, Middle -> RoundedCornerShape(16.dp)
-            else -> {
-                if (ownsMessage) ChatTheme.shapes.myMessageBubble else ChatTheme.shapes.otherMessageBubble
-            }
+    val messageBubbleShape = when (position) {
+        Top, Middle -> RoundedCornerShape(16.dp)
+        else -> {
+            if (ownsMessage) ChatTheme.shapes.myMessageBubble else ChatTheme.shapes.otherMessageBubble
         }
     }
 
-    val messageCardColor = when {
+    val messageBubbleColor = when {
         message.isGiphyEphemeral() -> ChatTheme.colors.giphyMessageBackground
         message.isDeleted() -> ChatTheme.colors.deletedMessagesBackground
         ownsMessage -> ChatTheme.colors.ownMessagesBackground
@@ -591,8 +587,8 @@ public fun DefaultMessageItemContent(
 
     MessageBubble(
         modifier = modifier,
-        shape = bubbleShape,
-        color = messageCardColor,
+        shape = messageBubbleShape,
+        color = messageBubbleColor,
         content = {
             when {
                 message.isGiphyEphemeral() -> {